feat: Implement 'list apis -e' Command
And update Docs
